Label close button
To ensure that users of assistive technologies, like screen readers, know what the close button is, we use `aria-label="Close"` to specify the accessible name.

Without it, the SVG would not be announced and assistive technology users would only know that there is a “button” without any indication of what that button does.

Signed-off-by: Eric Eggert <mail@yatil.net>